requirements.txt
tweepy
numpy
Pillow